Fix hot-fix and rollback handling
Added more documentation.

+++ update_db.sh
@@ -17,7 +17,8 @@ set -o errexit
 set -o pipefail
 
 FLYWAY_DOCKER_IMAGE=boxfuse/flyway:5.1.4-alpine
-SQL_FILE_NAME_PATTERN="TEST-<story_number>-<description_with_underscore_between_words>.<sql|rollback|hotfix>"
+SQL_FILE_NAME_PREFIX=TEST
+SQL_FILE_NAME_PATTERN="$SQL_FILE_NAME_PREFIX-<story_number>-<description_with_underscore_between_words>.<sql|rollback|hotfix>"
 
 enforce_mandatory_options() {
   # Checks for missing required options.
@@ -79,7 +80,7 @@ generate_schema_history() {
 	for version in "${hotfix_versions[@]}"; do
 		insert_batch_script+="insert into \"$SCHEMA\".schema_history(installed_rank, version, description, type, script,
 		installed_by, execution_time, success) select max(installed_rank) + 1, '$version', 'Hotfix version generation',
-		'PSQL', 'N/A', '$USERNAME', 0, true from schema_history;"
+		'SQL', 'N/A', '$USERNAME', 0, true from schema_history;"
 	done
 
 	log -i "Generating schema history"
@@ -89,21 +90,23 @@ generate_schema_history() {
 
 run_flyway_migration() {
 	log -i "Running Flyway migration using files with extensions '$1' and script folder $(pwd)/$SQL_FOLDER"
-	sudo docker run -v $(pwd)/"$SQL_FOLDER":/flyway/sql "$FLYWAY_DOCKER_IMAGE" -url="$URL" \
+	docker run --net=host -v $(pwd)/"$SQL_FOLDER":/flyway/sql "$FLYWAY_DOCKER_IMAGE" -url="$URL" \
 		-schemas="$SCHEMA" -user="$USERNAME" -password="$PASSWORD" -table=schema_history \
 		-baselineOnMigrate=true -baselineVersion=0 -ignoreMissingMigrations=true -outOfOrder=true \
-		-sqlMigrationPrefix=PHX- -sqlMigrationSeparator=- -sqlMigrationSuffixes="$1" -group=true migrate
+		-sqlMigrationPrefix="$SQL_FILE_NAME_PREFIX-" -sqlMigrationSeparator=- -sqlMigrationSuffixes="$1" -group=true migrate
 }
 
 delete_schema_history() {
 	declare -a versions=( $(find ./"$SQL_FOLDER" -name "*.$1" -type f | cut -d '-' -f 2) )
-	local quoted_versions_list=$(printf "'%s'," "${versions[@]}")
-	# Use string length to support older version of bash
-	quoted_versions_list="${quoted_versions_list:0:${#quoted_versions_list}-1}"
+	if [[ -n "${versions:-}" ]]; then
+		local quoted_versions_list=$(printf "'%s'," "${versions[@]}")
+		# Use string length to support older version of bash
+		quoted_versions_list="${quoted_versions_list:0:${#quoted_versions_list}-1}"
 
-	log i "Deleting schema history with versions $quoted_versions_list"
+		log i "Deleting schema history with versions $quoted_versions_list"
 
-	psql_execute_script "delete from \"$SCHEMA\".schema_history where version in($quoted_versions_list);"
+		psql_execute_script "delete from \"$SCHEMA\".schema_history where version in($quoted_versions_list);"
+	fi
 }
 
 psql_execute_script() {
@@ -126,7 +129,7 @@ rollback_changes() {
 
 print_version_state() {
 	log i "Printing '$SCHEMA' schema status"
-	sudo docker run "$FLYWAY_DOCKER_IMAGE" -url="$URL" -schemas="$SCHEMA" -user="$USERNAME" -password="$PASSWORD" \
+	docker run --net=host "$FLYWAY_DOCKER_IMAGE" -url="$URL" -schemas="$SCHEMA" -user="$USERNAME" -password="$PASSWORD" \
 		-table=schema_history info
 }
 
@@ -143,7 +146,7 @@ usage() {
         -p <password>
             Sets database password.
         -s <schema>
-            Sets database schema. This can be comma separated list, e.g.
+            Sets database schema.
         -f <sql_folder>
             Sets the folder where sql scripts are placed. Note that the script directory is used as root.
         -r Executes rollback scripts defined in <sql_folder>.
